[refactor]PostFormのRulesを別ファイルに分離
